% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoParApply.R
\name{autoParLapply}
\alias{autoParApply}
\alias{autoParLapply}
\alias{autoParSapply}
\title{*apply with auto-parallelization if available}
\usage{
autoParLapply(X, FUN, .maxCores = NULL, ...)

autoParSapply(X, FUN, .maxCores = NULL, ...)

autoParApply(X, MARGIN, FUN, .maxCores = NULL, ...)
}
\arguments{
\item{X, MARGIN, FUN, ...}{arguments to \code{\link{lapply}}, \code{\link{sapply}} and \code{\link{apply}} or their parallel equivalents}

\item{.maxCores}{limit the number of cores.}
}
\description{
autoPar*apply functions guess the number of cores to use and
run in parallel if possible. \dQuote{L}, \dQuote{S} and \dQuote{} (nothing) are available.
This function will make all objects in the global environment, and all currently loaded packages available to the parallel cluster nodes. This may fail in unexpected ways if other environments are \link[=attach]{attached} to the search path.
}
\examples{
autoParLapply(1:10, function(x) x^2)
autoParSapply(1:10, function(x) x^2)
x <- array(1:30, dim=c(2, 5, 3))
autoParApply(x, 1, sum)
autoParApply(x, c(1, 3), sum)
}

