% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcModifiedPeptideP.R
\name{calcModifiedPeptideP}
\alias{calcModifiedPeptideP}
\title{Calculate modified peptide probability}
\usage{
calcModifiedPeptideP(ModifiedSequences, PhosphoProbabilitySequences = NULL,
  OxProbabilitySequences = NULL)
}
\arguments{
\item{ModifiedSequences}{the "Modified Sequence" column in evidence.txt}

\item{PhosphoProbabilitySequences}{the "Phospho (STY) Probabilities" column in evidence.txt}

\item{OxProbabilitySequences}{the "Oxidation (M) Probabilities" column in evidence.txt}
}
\value{
the numeric probabilities
}
\description{
Integrates the Modification Probabilities from MaxQuant into a localization probability of the peptide
}
\examples{
calcModifiedPeptideP("_(ac)AGDS(ph)DSWDADAFSVEDPVRK_", "AGDS(1)DSWDADAFSVEDPVRK", "")
calcModifiedPeptideP("_AAFNSGKVDIVAINDPFIDLNYM(ox)VYM(ox)FQYDSTHGK_", "",
                     "AAFNSGKVDIVAINDPFIDLNYM(1)VYM(1)FQYDSTHGK")
calcModifiedPeptideP("_AAEM(ox)CY(ph)RK_", "AAEMCY(1)RK", "AAEM(1)CYRK")

# Also vectorized:
calcModifiedPeptideP(
                     c("_AAEM(ox)CY(ph)RK_", "_(ac)AGDS(ph)DSWDADAFSVEDPVRS(ph)_",
                       "_(ac)AGDS(ph)DSWDADAFSVEDPVRM(ox)_"),
                     c("AAEMCY(0.99)RK", "AGDS(0.995)DSWDADAFS(0.007)VEDPVRS(0.998)",
                       "AGDS(0.995)DSWDADAFS(0.007)VEDPVRM"),
                     c("AAEM(0.87)CYRK", "", "AGDSDSWDADAFSVEDPVRM(0.998)"))
}
