% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constructModifiedPeptide}
\alias{constructModifiedPeptide}
\title{Construct the modified peptide sequence as a string}
\usage{
constructModifiedPeptide(ModifiedSequences, positions = integer(),
  delimiter = ";")
}
\arguments{
\item{ModifiedSequences}{the "Modified Sequence" column in evidence.txt}

\item{positions}{the peptides start positions within the protein. Recycled according to the standard rules if necessary, with a warning.}

\item{delimiter}{an optional delimiter to separate the modifications within one peptide}
}
\value{
the modifications string
}
\description{
Takes the Modified Sequence from MaxQuant and turns it into a string of the form a;T_12;M_72.
}
\examples{
constructModifiedPeptide("_(ac)AGDS(ph)DSWDADAFSVEDPVRK_", 1)
constructModifiedPeptide("_AAFNSGKVDIVAINDPFIDLNYM(ox)VYM(ox)FQYDSTHGK_", 20)
constructModifiedPeptide("_AAEM(ox)CY(ph)RK_", 10, ":")

# Also vectorized:
constructModifiedPeptide(c("_(ac)AGDS(ph)DSWDADAFSVEDPVRK_",
                       "_AAFNSGKVDIVAINDPFIDLNYM(ox)VYM(ox)FQYDSTHGK_",
                       "_AAEM(ox)CY(ph)RK_"), c(1, 20, 10), "+")
}
