% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elongate.R
\name{elongate}
\alias{elongate}
\title{Elongates a pseudo-wide column}
\usage{
elongate(id, wide, pattern = ";")
}
\arguments{
\item{id}{a column that will not be splitted}

\item{wide}{the column to split}

\item{pattern}{the character or regex used to split up the wide column (see \link{str_split})}
}
\description{
Takes a pseudo-wide column and tranform it into a long format
}
\examples{
data(pseudo.wide)
elongated.score <- elongate(pseudo.wide$Protein, pseudo.wide$Score)
elongated.sequence <- elongate(pseudo.wide$Protein, pseudo.wide$Sequence)
# The following works only if the Score and Sequence have the same pattern
# (ie same number of splits per ID)
elongated <- data.frame(Protein = elongated.score$id, 
                        Score = elongated.score$wide,
                        Sequence = elongated.sequence$wide)
\dontshow{
stopifnot(identical(dim(elongated.score), dim(elongated.sequence)))
}
}

