% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMart.R
\name{getMart}
\alias{getMart}
\title{Get the Mart corresponding the Ensembl version and species}
\usage{
getMart(version, species = "human")
}
\arguments{
\item{version}{the ensembl version (number) or \dQuote{current}}

\item{species}{the species, currently either \dQuote{human} or \dQuote{mouse} in English, or the species part of the dataset name (without the \dQuote{_gene_ensembl} part, as listed by \code{\link[biomaRt]{listDatasets}}, for instance \dQuote{rnorvegicus} or \dQuote{scerevisiae}).}
}
\description{
Get the Mart corresponding the Ensembl version and species
}
\examples{
if(require("biomaRt")) {
  ensembl.proteins <- c("ENSP00000361930", "ENSP00000309503")
  mart <- getMart(88, "human")
  getBM(attributes = c("ensembl_peptide_id", "uniprotswissprot", "refseq_peptide", "hgnc_symbol"), filters = "ensembl_peptide_id", values = ensembl.proteins, mart = mart)

  # Mouse example
  ensembl.proteins <- c("ENSMUSP00000004055", "ENSMUSP00000002708")
  mart <- getMart(88, "mmusculus")
  getBM(attributes = c("ensembl_peptide_id", "uniprotswissprot", "refseq_peptide", "mgi_symbol"), filters = "ensembl_peptide_id", values = ensembl.proteins, mart = mart)
}
}
