% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal.R
\name{\%==\%}
\alias{\%==\%}
\alias{equal}
\alias{equality}
\title{Test for near equality of two vectors}
\usage{
x \%==\% y
}
\arguments{
\item{x, y}{atomic vectors}
}
\description{
This function is a vectorized version of \code{\link{all.equal}}. Instead of checking if all elements are nearly equal, it does the comparison element-wise like \code{==}. Arguments are recycled according to the standard rules.
}
\examples{
a <- c(0.1, 0.2, 0.3) + 0.1
b <- c(0.2, 0.3, 0.4)
a == b
a \%==\% b
}
