% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concatenate.R
\name{\%+\%}
\alias{\%+\%}
\title{Concatenate two character strings without any whitespace.}
\usage{
x \%+\% y
}
\arguments{
\item{x,y}{the strings to concatenate}
}
\value{
the concatenated string
}
\description{
This is strictly equivalent to a call to paste0, with a bit less typing, similar to many other progamming languages that use \code{+} (or sometimes \code{.}) to concatenate strings. This function will take any type of vector happily, but make sense only with character vectors of length 1.
}
\examples{
"Hello" \%+\% " World!"
}

