% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessCores.R
\name{guessCores}
\alias{guessCores}
\title{Guesses a number of CPU to use for parallel calls
Especially useful for use with \code{\link{makeCluster}}.}
\usage{
guessCores(verbose = TRUE)
}
\arguments{
\item{verbose}{show the number of cores that was guessed}
}
\value{
a number of cores to use
}
\description{
First checks the environment variable NCPUS. This will typically be set in PBS jobs to the number of cores assigned to the job.
Otherwise, will use the number of CPUs with \code{\link{detectCores}}. If that fails or returns more than 12 cores, it will go back to 1 (as this is indicative that the job is being run on a supercomputer but without the NCPUS properly set)
}
\examples{
print(ncpus <- guessCores())
# With the parallel package
library(parallel)
cl = makeCluster(ncpus)
parLapply(cl = cl, 1:ncpus, log)
stopCluster(cl)
}

