% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lettersToDecimal.R
\name{lettersToDecimal}
\alias{lettersToDecimal}
\title{Converts a letters-based representation into normal base 10 representation}
\usage{
lettersToDecimal(..., space = letters)
}
\arguments{
\item{...}{objects to be converted}

\item{space}{the original space, such as \code{\link{letters}}, \code{\link{LETTERS}}  or \code{1:10} (which will do nothing)}
}
\description{
In effect, this reverses the effect of \href{https://linux.die.net/man/1/split}{split} where output files are numbered with letters. This can be used to convert from other spaces too.
}
\examples{
lettersToDecimal("aaaa")
lettersToDecimal("bczf")

# Using the space of uppercase letters
lettersToDecimal("BCFZ", space = LETTERS)

# To work with mixed case, use tolower or similar
lettersToDecimal(tolower("Bczf"))

# Multiple inputs, the following lines are equivalent:
lettersToDecimal("aaaa", "bczf")
lettersToDecimal(c("aaaa", "bczf"))
}
