% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadToList.R
\name{loadTo}
\alias{loadTo}
\alias{loadToList}
\alias{loadToEnv}
\title{Load to a list of environment
Loads an R dataset written with the \code{\link{save}} function into a \code{\link{list}} or \code{\link{environment}}}
\usage{
loadToList(file)

loadToEnv(file)
}
\arguments{
\item{file}{the \code{\link{file}} argument for \code{\link{load}}, see there for more details}
}
\value{
a list or environment with the named contents of \code{\link{file}}
}
\description{
Load to a list of environment
Loads an R dataset written with the \code{\link{save}} function into a \code{\link{list}} or \code{\link{environment}}
}
\examples{
fname <- tempfile()
a <- "b"
b <- 5
save(a, b, file=fname)
l <- loadToList(fname)
\dontshow{
stopifnot(identical(l, list(a="b", b=5)))
}
e <- loadToEnv(fname)
\dontshow{
testEnv <- new.env()
assign("a", "b", envir = testEnv)
assign("b", 5, envir = testEnv)
# Can't use identical on environment
stopifnot(isTRUE(all.equal(e, testEnv)))
}
file.remove(fname)
}
