% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadToList.R
\name{loadToList}
\alias{loadToList}
\title{Load an R dataset written with the \code{\link{save}} function into a \code{\link{list}}}
\usage{
loadToList(file)
}
\arguments{
\item{file}{the \code{\link{file}} argument for \code{\link{load}}, see there for more details}
}
\value{
a list with the named contents of \code{\link{file}}
}
\description{
Load an R dataset written with the \code{\link{save}} function into a \code{\link{list}}
}

