% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktemp.R
\name{mktemp}
\alias{mktemp}
\title{Create a temporary file securely}
\usage{
mktemp(template = "file.XXXXXXXXXX", tmpdir = tempdir())
}
\arguments{
\item{template}{the template, which must end with at least 3 'X's. Only alphanumeric characters, underscore and dots are allowed.}

\item{tmpdir}{the containing temporary folder. Only alphanumeric characters, underscore and dots are allowed.}
}
\value{
the newly created file name
}
\description{
Creates a temporary file securely with a system call to mktemp. A version of mktemp from GNU coreutils or BSD is required.
}
\details{
R's \code{\link{tempfile}} function makes no guarantee against race conditions, where the same temp file could be generated by different R processes. This is a problem especially 26 *in parallel environments. By contrast, file names generated by \code{mktemp} are guaranteed to be unique.

Note that the file is created before the function returns and must be removed manually afterwards.
}
\examples{
tmpfile <- mktemp()
file.remove(tmpfile)
}

