% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.warn.R
\name{na.warn}
\alias{na.warn}
\title{Remove missing values with a warning}
\usage{
na.warn(object, ...)
}
\arguments{
\item{object}{an R object}

\item{...}{additional arguments}
}
\description{
This function behaves exactly like \code{\link{na.omit}}, except that it prints a warning if any missing value was removed
}
\examples{
DF <- data.frame(x = c(1, 2, 3), y = c(0, 10, NA))
na.warn(DF)
}
