% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.wide.R
\docType{data}
\name{pseudo.wide}
\alias{pseudo.wide}
\title{A dummy data set in pseudo-wide format}
\format{a \code{data.frame} with 3 columns.}
\description{
This data set list proteins with two pseudo-wide columns
containing the identified sequence and associate score
}

