% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{rank.normalize}
\alias{rank.normalize}
\title{Rank normalization
This function fits the data exactly to the normal distribution (or any other quantile function given as \code{FUN}) using corresponding quantiles.}
\usage{
rank.normalize(x, FUN = qnorm, ties.method = "average", na.action)
}
\arguments{
\item{x}{the data to normalize}

\item{FUN}{a quantile function such as \code{\link{qnorm}}}

\item{ties.method}{the \code{ties.method} argument for \code{\link{rank}}}

\item{na.action}{a function to handle \code{NA}s. Defaults to the value of \code{\link{options}("na.action")}}
}
\value{
the normalized data
}
\description{
Rank normalization
This function fits the data exactly to the normal distribution (or any other quantile function given as \code{FUN}) using corresponding quantiles.
}
\examples{
not.normal <- rlnorm(100)
hist(not.normal)
hist(rank.normalize(not.normal))

# Different function
hist(rank.normalize(not.normal, FUN = qexp))

# Breaking ties
hist(rank.normalize(not.normal, ties.method = "random"))

# Missing values
not.normal[10] <- NA
hist(rank.normalize(not.normal)) # Shouldn't fail
\dontrun{hist(rank.normalize(not.normal, na.action = na.fail))} # Should fail

}

