% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe.mapping.R
\name{safe.mapping}
\alias{safe.mapping}
\title{Maps with a left_join, ensuring constant number of rows}
\usage{
safe.mapping(x, map, by = NULL)
}
\arguments{
\item{x}{the \code{\link{data.frame}} with fixed size}

\item{map}{a \code{\link{data.frame}} with the new columns.}

\item{by}{see \code{\link{left_join}}}
}
\description{
This avoids generating left duplicates if entries are duplicated in the right \code{by}.
If duplicates are generated, the function will stop with an error.
}
\examples{
a <- data.frame(A = LETTERS[1:5], B = 1:5)
b <- data.frame(A = LETTERS[1:5], C = 15:11)
dplyr::left_join(a, b)
safe.mapping(a, b)

# Now with duplicates
\dontrun{
c <- data.frame(A = LETTERS[1:5], C = 20:11)
dplyr::left_join(a, c) # creates duplicates of x
safe.mapping(a, c) # stops with an error
}

}
