% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff_symmetric.R
\name{setdiff_symmetric}
\alias{setdiff_symmetric}
\alias{sets}
\title{Symmetric setdiff}
\usage{
setdiff_symmetric(x, y)
}
\arguments{
\item{x, y}{objects to be compared}
}
\description{
performs \emph{symmetric} set difference
}
\details{
The \code{\link[base:sets]{setdiff}} function from the \pkg{base} package performs assymetric difference of its arguments and returns the members of \code{x} missing in \code{y}. This function performs a symmetric difference instead, where both members of \code{x} missing in \code{y} and members of \code{y} missing in \code{x} will be returned.

If \pkg{dplyr} is present, \code{\link{data.frame}}s can also be used as input.
}
\examples{
setdiff_symmetric(1:10, 5:15)
}
