% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformations}
\alias{FUN_transform}
\alias{log_transform}
\alias{power_transform}
\alias{sqrt_transform}
\alias{transform}
\alias{transformations}
\title{Transformations}
\usage{
sqrt_transform(x)

power_transform(x, exp = 1/2)

log_transform(x, base = exp(1))

FUN_transform(x, FUN, ...)
}
\arguments{
\item{x}{the data vector to transform}

\item{exp}{the exponent for the transoformation x^exp}

\item{base}{the logarithm base}

\item{FUN}{the transformation function, defined on the range [0, +Inf]}

\item{...}{additional arguments for \code{FUN}}
}
\description{
Various scale transformation functions. These will keep the sign of x (and are basically \code{sign(x) * FUN(abs(x))})
}
\examples{
curve(sqrt_transform(x), -2, 2, n = 1000)
curve(power_transform(x), -2, 2, n = 1000)
curve(power_transform(x, 3), -2, 2, n = 1000)
curve(log_transform(x), -2, 2, n = 1000)
curve(log_transform(x, 10), -2, 2, n = 1000)
curve(FUN_transform(x, FUN = function(x) x + 2 * x^3), -2, 2)
curve(FUN_transform(x, FUN = function(x, a, b) x + a * x^b, a = 2, b = 3), -2, 2)
curve(FUN_transform(x, FUN = function(x, a, b) x + a * x^b, a = 4, b = 5), -2, 2)
}

