\name{xavamess-package}
\alias{xavamess-package}
\alias{xavamess}
\docType{package}
\title{
Various useful functions
}
\description{
This package contains one useful function
}
\details{
\tabular{ll}{
Package: \tab xavamess\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-02-05\cr
License: \tab GPL-3\cr
}
}
\author{
Xavier Robin

Maintainer: Xavier Robin <xavier.robin@bric.ku.dk>
}
%\references{
%~~ Literature or other references for background information ~~
%}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
