% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoParApply.R
\name{autoParLapply}
\alias{autoParApply}
\alias{autoParLapply}
\alias{autoParSapply}
\title{*apply with auto-parallelization if available
autoPar*apply functions guess the number of cores to use and
run in parallel if possible.}
\usage{
autoParLapply(X, FUN, ...)

autoParSapply(X, FUN, ...)

autoParApply(X, MARGIN, FUN, ...)
}
\arguments{
\item{X, FUN, ...}{arguments to \code{\link{lapply}}}
}
\description{
*apply with auto-parallelization if available
autoPar*apply functions guess the number of cores to use and
run in parallel if possible.
}
\examples{
autoParLapply(1:10, function(x) x^2)
autoParSapply(1:10, function(x) x^2)
x <- array(1:30, dim=c(2, 5, 3))
autoParApply(x, 1, sum)
autoParApply(x, c(1, 3), sum)
}

