% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{sqrt.transform}
\alias{log.transform}
\alias{power.transform}
\alias{sqrt.transform}
\title{Transfomrations}
\usage{
sqrt.transform(x)

power.transform(x, exp = 1/2)

log.transform(x, base = exp(1))
}
\arguments{
\item{x}{the data to transform}

\item{exp}{the exponent for the transoformation x^exp}

\item{base}{the logarithm base}
}
\description{
Various scale transformation functions. These will keep the sign of x (and are basically \code{sign(x) * FUN(abs(x))})
}
\examples{
curve(sqrt.transform(x), -2, 2, n = 1000)
curve(power.transform(x), -2, 2, n = 1000)
curve(power.transform(x, 3), -2, 2, n = 1000)
curve(log.transform(x), -2, 2, n = 1000)
curve(log.transform(x, 10), -2, 2, n = 1000)
}

